Semrush integration (#4)
* openai integration

* openai: remove unused model

* refactor controller package

* Semrush integration

* Semrush integration, more APIs

* adding keyword analytic API

* Adding unit tests

* update traffic API

---------

+++ src/main/java/com/searchify/suggestion/entity/semrush/response/SemrushTopSubdomainResponse.java
@@ -26,7 +26,7 @@ public class SemrushTopSubdomainResponse extends SemrushBaseResponse {
     private LocalDate displayDate;
 
     @CsvBindByName(column = "total_visits")
-    private Integer totalVisits;
+    private Long totalVisits;
 
     @CsvBindByName(column = "desktop_share")
     private Double desktopShare;
